Radar added

+++ Assets/Scripts/GPS Tracking/Radar/MarkerScript.cs
@@ -18,7 +18,7 @@ public class MarkerScript : MonoBehaviour
 
     public void DrawMarker(float maxDistance, int length)
     {
-        var angle = -((float)MathHelper.GetAzimuth(GPSTracker.Instance.CurrentLocation, LocationProvider.Location) + Mathf.PI / 2);
+        var angle = -(MathHelper.GetAzimuth(GPSTracker.Instance.CurrentLocation, LocationProvider.Location) + Mathf.PI / 2);
         var distance = Location.DistanceBetween(GPSTracker.Instance.CurrentLocation, LocationProvider.Location);
 
         if (distance > maxDistance)
